[api] easier external numeric types #1394
- add vd.numericTypes mutable list
- add @vd.numericType() decorator for external types
- refactor floatsi into visidata.type_floatsi
- refactor date into visidata.type_date
- refactor currency into visidata.type_currency

Now you can make e.g. `log10` a valid numeric type for a column:

    @vd.numericType('/')
    def log10(*args):
        if not args:
            return 0
        return math.log10(*args) if args[0] > 0 else -10

    Sheet.addCommand('', 'type-log10', 'cursorCol.type = log10')

which might make it nicer to work with.
